pyproject.toml
[project]
name = "autogen-101"
version = "0.1.0"
description = "This is a project to learn Autogen with hands-on examples"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ag2[openai]>=0.9",
    "ipykernel>=6.29.5",
    "nbformat>=5.10.4",
    "pyautogen>=0.9",
    "python-dotenv>=1.1.0",
]
